ascii/requirements.txt
Pillow
urllib3